fix(x86): define CR0, CR4, and EFER bitflags as u64

+++ alioth/src/arch/x86_64/msr.rs
@@ -29,7 +29,7 @@ pub const IA32_TSC_AUX: u32 = 0xc000_0103;
 pub const IA32_MISC_ENABLE: u32 = 0x0000_01a0;
 
 bitflags! {
-    pub struct Efer(u32) {
+    pub struct Efer(u64) {
         /// SYSCALL enable
         SCE = 1 << 0;
         /// IA-32e mode enable
